refactor(toolbar): Revamp toolbar to use toggle group
- Replace dropdown menus in the toolbar with a toggle group for
  selecting tool types, enhancing usability and providing a cleaner
  interface.
- Implement `ToggleGroup` from Radix UI to manage tool selections,
  allowing single selection mode for cleaner visual feedback.
- Icons for `addTextNode`, `addFileNode`, and `addTextGenerationNode`
  are now directly represented in the toolbar for better accessibility.

This change simplifies the tool selection process and improves
the overall user experience within the toolbar.

No breaking changes.
