+++ package.json
@@ -43,6 +43,7 @@
 		"@radix-ui/react-slider": "1.2.1",
 		"@radix-ui/react-slot": "^1.1.0",
 		"@radix-ui/react-tabs": "^1.1.0",
+		"@radix-ui/react-toggle-group": "1.1.0",
 		"@radix-ui/react-tooltip": "1.1.4",
 		"@sentry/nextjs": "8.34.0",
 		"@sentry/wizard": "^3.26.0",
